apps/transaction-manager/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
databricks-sdk
databricks-sql-connector
